build: handle diffrent input between forms as childrens + convert check car modal to page
