+++ client/src/assets/data/features.ts
@@ -12,7 +12,7 @@ const features :any = {
   'Steering wheel control': 'https://carswitch.com/assets/png/14.png',
   'Auto dimming mirror': 'https://carswitch.com/assets/png/34.png',
   'Headsup display': 'https://carswitch.com/assets/png/35.png',
-  'Birds eye view camera': 'https://carswitch.com/assets/png/33.png',
+  'Birdseye view camera': 'https://carswitch.com/assets/png/33.png',
   'Rear mirror camera': 'https://carswitch.com/assets/png/09.png',
   'Adaptive cruis control': 'https://carswitch.com/assets/png/01.png',
   'Digital display': 'https://carswitch.com/assets/png/04.png',
@@ -61,7 +61,7 @@ export const featuresArray :any = ['Power seats',
   'Steering wheel control',
   'Auto dimming mirror',
   'Headsup display',
-  'Birds eye view camera',
+  'Birdseye view camera',
   'Rear mirror camera',
   'Adaptive cruis control',
   'Digital display',
