+++ client/src/index.css
@@ -2,7 +2,7 @@
 @import './cssReset.css';
 
 :root {
-  --main-color: #0A20E6;
+  --main-color: #72a0fc;
   --text-color: #2f3643;
   --second-text-color: #A6AFBA;
   --container-color: #FFF;
